Prototype for doing concurrency protection in PlayUAV.
This is the literal, rote way to do it. I am also thinking about trying out a set of helper functions for get/set -- it's going to be a lot of typing either way, and I think the latter may prove easier to read after the fact, and to think about. Worth trying next anyhow.