More concurrency protection.
This time I tried using wrapper/accessor functions to insulate us from the shared variable, and to make it very hard to accidentally use the raw variable. I started with the GPS coordinates, which should be updated together and are arguably the most complex use case for the mutexes, so I wanted to get it out of the way.

This compiles but is untested. Assuming it works, from here we'd move through the various variables in a most-to-least important order. osd_heading is next, unsure what comes after that.