This is overall promising!
Bad: On startup we get a 45 second locked-on-a-mutex-feeling problem where the screen is half drawn and the VSP task it down. Then, DONG! and USB is up and the two loops are running smoothly. So, it seems like it must be a fixable problem - some startup issue, then a timeout? But what? Need to test that next.

Very good: The loop is running smoothly after the timeout on alt & lat & lon, with the mutexes in full effect! So, that's great, and it does seem correct to me. The airlock is protected, yet both sides of the loop can run smoothly without excessive contention on it.