- Fixing error in display of GPS error message pointed out by Paul A here: http://fpvlab.com/forums/showthread.php?40658-A-Graphics-OSD-for-APM-Pixhawk&p=814872&viewfull=1#post814872
He observed that the "NO GPS FIX" error should clear once there is a 3D lock, and it does not. He's was right.

In the process of fixing this, I corrected a number of other problems with the draw_warning() routine and associated code. Now it does what I believe it was originally supposed to do: Display each active error message in turn for a fixed number of (now configurable) milliseconds, as long as there are any errors. When there are no errors, it displays nothing.

I would like to take preflight error reporting a bit farther, but this is just a bug fix, attempting to get the code stable.

- I also found an error I'd already fixed creeping back in on the display of "H -M", showed "H null". This was from a bad cross-branch merge during pull request creation, my apologies.

- Bumping descriptive string for build.